Clean up msg.cpp (#502)
* Clean up msg_wait_for_turns
* Clean up msg_get_next_packet
* Clean up msg_pre_packet
* Clean up DeltaSaveLevel
* Simple clean ups for msg.cpp
* Clean up NetSendCmdQuest
* Clean up NetSendCmdGItem
* Clean up NetSendCmdGItem2
* Clean up NetSendCmdPItem
* Clean up NetSendCmdChItem
* Clean up NetSendCmdDItem
* Clean up multi_send_packet
* Clean up msg_errorf
* Clean up i_own_level
* Clean up msg_free_packets

+++ Source/msg.h
@@ -41,7 +41,7 @@ void __fastcall NetSendCmdParam1(BOOL bHiPri, BYTE bCmd, WORD wParam1);
 void __fastcall NetSendCmdParam2(BOOL bHiPri, BYTE bCmd, WORD wParam1, WORD wParam2);
 void __fastcall NetSendCmdParam3(BOOL bHiPri, BYTE bCmd, WORD wParam1, WORD wParam2, WORD wParam3);
 void __fastcall NetSendCmdQuest(BOOL bHiPri, BYTE q);
-void __fastcall NetSendCmdGItem(BOOL bHiPri, BYTE bCmd, BYTE mast, BYTE pnum, int ii);
+void __fastcall NetSendCmdGItem(BOOL bHiPri, BYTE bCmd, BYTE mast, BYTE pnum, BYTE ii);
 void __fastcall NetSendCmdGItem2(BOOL usonly, BYTE bCmd, BYTE mast, BYTE pnum, TCmdGItem *p);
 BOOL __fastcall NetSendCmdReq2(BYTE bCmd, BYTE mast, BYTE pnum, TCmdGItem *p);
 void __fastcall NetSendCmdExtra(TCmdGItem *p);
